Camera occluder/Kit Hiding. round 3
Re-structuring with pre-assigned "facings." Will need to change the elevation controller to reference facings and not kit objects if this ends up working. Probably for the best, it will reduce the amount of assets running coroutines by a significant amount. In progress. Pulling this will break the occluder and the elevation control