Elevation controller update to work with facings
Facings updated to work with elevation controller and ready to be implemented with occlusion controller. Facings collect all kit asset children and start a coroutine to control all of them with an alpha over time controlling a scale lerp. This has a bool arguement to hide the facing entirely when finished or to leave the bottom slice. This will sllow it to remain partially visible when called from the camera occluder or entirely hidden when called from the elevation controller