Occluder progress
Occluder round 2. Need to remove reference to toggle as it has no awareness of elevation controller. Needs absolute "on-Off" commands. Also, direction for occluder should be camera forward to distance from target. Not direction to target. Then as long as the camera moves to focus on unit it will be fine and more predictable.