Occlusion with overlap sphere.
Reliable. Consistent. Still not really what I want but checking it in so I can pull it back If I break it. Overlap sphere works well but I want to try a raycast to the navcursor on nav layer so only objects between get hidden. Currenlty all in radious get captured regarldess of their location relative to cursor.